% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_desc_file.R
\name{check_desc_file}
\alias{check_desc_file}
\title{Check \code{DESCRIPTION} File for Required Fields}
\usage{
check_desc_file(file, verbose = TRUE)
}
\arguments{
\item{file}{A character string specifying the path to the \code{DESCRIPTION} file.}

\item{verbose}{A logical parameter; if set to TRUE, detailed messages will
be displayed during the check.}
}
\value{
A logical value; \code{TRUE} if all required fields are present, \code{FALSE}
otherwise.
}
\description{
This function ensures that a \code{DESCRIPTION} file contains all of the necessary
fields for an R package. It operates verbosely by default, providing alerts
about the checking process.
}
