% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_pkg.R
\name{check_description}
\alias{check_description}
\title{Check fields in DESCRIPTION file}
\usage{
check_description(path)
}
\arguments{
\item{path}{path to DESCRIPTION file}
}
\value{
logical
}
\description{
Check fields in DESCRIPTION file
}
\examples{
# check_description("path/to/DESCRIPTION")
}
