% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_pkg.R
\name{check_lines}
\alias{check_lines}
\title{Check the lines in text file}
\usage{
check_lines(file, regex)
}
\arguments{
\item{file}{path to file}

\item{regex}{}
}
\value{
logical value from \code{grep()}
}
\description{
Check the lines in text file
}
\examples{
# check_lines("DESCRIPTION", "Package:")
}
