% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_pkg.R
\name{check_pkg_rproj}
\alias{check_pkg_rproj}
\title{Check package fields in .Rproj file}
\usage{
check_pkg_rproj(path, verbose = FALSE)
}
\arguments{
\item{path}{path to .Rproj file}
}
\value{
logical
}
\description{
Check package fields in .Rproj file
}
\examples{
# check_pkg_rproj(getwd())
}
