% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_r_pkg.R
\name{check_r_pkg}
\alias{check_r_pkg}
\title{Check Directory for R Package Structure}
\usage{
check_r_pkg(path, verbose = FALSE)
}
\arguments{
\item{path}{A character string specifying the path to the directory to be
checked.}

\item{verbose}{A logical parameter; if set to TRUE, detailed messages will
be displayed during the check.}
}
\value{
A logical value; \code{TRUE} if both \code{.Rproj} and \code{DESCRIPTION} files
are configured correctly, \code{FALSE} otherwise.
}
\description{
This function checks a directory to ascertain if it has the required files
(\code{.Rproj} and \code{DESCRIPTION}) configured correctly for an R package. It
provides verbose output detailing which checks have passed or failed.
}
\section{Details}{

This function wraps around \code{check_desc_file} and \code{check_rproj_file}
functions to check both the \code{DESCRIPTION} and \code{.Rproj} files. If \code{verbose}
is \code{TRUE}, it will use \code{cli} package functions to alert the user about the
checks being performed and their results.
}

\examples{
check_r_pkg(
     system.file("pkg", package = "shinyap"),
     verbose = TRUE)
check_r_pkg(
     system.file("app", package = "shinyap"),
     verbose = TRUE)
}
