% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_pkg.R
\name{check_rproj}
\alias{check_rproj}
\title{Check fields in .Rproj file}
\usage{
check_rproj(rproj_file_path)
}
\arguments{
\item{path}{path to .Rproj file}
}
\value{
logical
}
\description{
Check fields in .Rproj file
}
\examples{
# check_rproj("path/to/file.Rproj")
}
