% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rproj_file.R
\name{check_rproj_file}
\alias{check_rproj_file}
\title{Check .Rproj File for Required Fields}
\usage{
check_rproj_file(file, verbose = FALSE)
}
\arguments{
\item{file}{character string specifying the path to the \code{.Rproj} file.}

\item{verbose}{logical; if  \code{TRUE}, detailed messages will be displayed}
}
\value{
A logical value; \code{TRUE} if all fields are found, \code{FALSE} otherwise.
}
\description{
Check .Rproj File for Required Fields
}
\examples{
# check_rproj_file("path/to/project.Rproj", verbose = TRUE)
}
