% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_text_field.R
\name{check_txt_field}
\alias{check_txt_field}
\title{Check for fields in text file}
\usage{
check_txt_field(file, field)
}
\arguments{
\item{file}{path to plain text file}

\item{field}{field in file}
}
\value{
logical
}
\description{
Check for fields in text file
}
