% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_app.R
\name{clone_app}
\alias{clone_app}
\title{Clone application from Shiny App-Packages}
\usage{
clone_app(app = "main", open = TRUE)
}
\arguments{
\item{app}{name of app}

\item{open}{open app in new Posit Workbench session}

\item{...}{additional arguments passed to \code{gert::git_clone()}}
}
\value{
shiny app (or app-package) branch
}
\description{
Clone application from Shiny App-Packages
}
\examples{
# clone_app("02_movies-app")
}
