% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_pkg.R
\name{is_r_pkg}
\alias{is_r_pkg}
\title{Is directory an R package?}
\usage{
is_r_pkg(path = getwd(), verbose = TRUE)
}
\arguments{
\item{path}{path to folder}

\item{verbose}{verbose messages? Defaults to \code{TRUE}}
}
\value{
logical
}
\description{
Is directory an R package?
}
\examples{
# is_r_pkg()
}
