% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch shiny app}
\usage{
launch_app(app)
}
\arguments{
\item{app}{application branch}
}
\value{
shiny app
}
\description{
Launch shiny app
}
\examples{
# launch_app("02_movies-app")
# launch_app("07_roxygen2")
}
