% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Inverted version of \code{\%in\%}}
\usage{
x \%nin\% table
}
\description{
Returns \code{TRUE} for elements of \code{x} that are \emph{not} in \code{table}.
}
\examples{
1 \%nin\% 1:10
"A" \%nin\% 1:10
}
\keyword{internal}
