% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_text_field.R
\name{check_text_field}
\alias{check_text_field}
\title{Check for fields in text file}
\usage{
check_text_field(file, field, verbose = FALSE)
}
\arguments{
\item{file}{path to plain text file}

\item{field}{field in file}
}
\value{
logical
}
\description{
Check for fields in text file
}
\examples{
check_text_field("Package: shinypak \n Title: Getting your app into an R package",
"Package")
check_text_field("Package: shinypak \n Title: Getting your app into an R package",
"Package", verbose = TRUE)
}
