% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_package.R
\name{is_r_package}
\alias{is_r_package}
\title{Check Directory for R Package Structure}
\usage{
is_r_package(path, verbose = FALSE)
}
\arguments{
\item{path}{A character string specifying the path to the directory to be
checked.}

\item{verbose}{A logical parameter; if set to TRUE, detailed messages will
be displayed during the check.}
}
\value{
A logical value; \code{TRUE} if both \code{.Rproj} and \code{DESCRIPTION} files
are configured correctly, \code{FALSE} otherwise.
}
\description{
This function checks a directory to ascertain if it has the required files
(\code{.Rproj} and \code{DESCRIPTION}) configured correctly for an R package. It
provides verbose output detailing which checks have passed or failed.
}
\section{Details}{

This function wraps around \code{is_pkg_description} and \code{is_pkg_rproj}
functions to check both the \code{DESCRIPTION} and \code{.Rproj} files. If \code{verbose}
is \code{TRUE}, it will use \code{cli} package functions to alert the user about the
checks being performed and their results.
}

\examples{
is_r_package(
     system.file("pkg", package = "shinyap"),
     verbose = TRUE)
is_r_package(
     system.file("app", package = "shinyap"),
     verbose = TRUE)
}
