% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch shiny app}
\usage{
launch(app)
}
\arguments{
\item{app}{application branch}
}
\value{
shiny app
}
\description{
Launch shiny app
}
\examples{
# launch("02_movies-app")
# launch("07_roxygen2")
}
