% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_apps.R
\name{list_apps}
\alias{list_apps}
\title{List Branches of the sap Repository}
\usage{
list_apps(regex = NULL)
}
\arguments{
\item{regex}{An optional character string containing a regular expression
(regex) to filter branch names}
}
\value{
A data frame with two columns: \code{branch} (the branch name) and
\code{last_updated} (the date the branch was last updated), ordered
alphabetically by branch name. Branches named \code{HEAD}, \code{main}, and
\code{gh-pages} are excluded.
}
\description{
This function clones the 'sap' GitHub repository, extracts the list
of branches along with their last update time, and then returns this information
as a tibble. The repository is cloned to a temporary directory, which is removed
after the function executes.
}
\section{How it works}{

The function operates in several steps:
\enumerate{
\item It saves the current working directory to revert back to it later
\item It clones the \href{https://github.com/mjfrigaard/sap}{'sap'}
repository from GitHub into a temporary directory
\item It lists all branches of the cloned repository
\item It separates the branch \code{name}s into \code{source} and \code{branch_name}
\item If \code{regex} is supplied, it filters branches to those matching the pattern
\item It cleans up by deleting the temporary directory and restoring the original
working directory
}
}

\examples{
list_apps()
list_apps(regex = "tests")
}
