% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_apps.R
\name{list_apps}
\alias{list_apps}
\title{List available app packages}
\usage{
list_apps(regex)
}
\arguments{
\item{regex}{passed to list.files(pattern)}
}
\value{
tibble of available app-packages
}
\description{
List available app packages
}
\examples{
list_app_pkgs()
list_app_pkgs(regex = "tests")
}
