% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_cols.R
\name{sep_cols}
\alias{sep_cols}
\title{Separate columns (base)}
\usage{
sep_cols(data, col, into, sep)
}
\arguments{
\item{data}{data.frame or tibble}

\item{col}{column to split}

\item{into}{columns to split into}

\item{sep}{regex to split on}
}
\value{
data with new columns
}
\description{
Separate columns (base)
}
\keyword{internal}
